Adding Histology data + Minor update to code
One .xlsx file (CellCounts_Histology) was added countaining data regarding
density of GABAergic neurons in the hilus of the dentate gyrus and in the
granule cell layers.

The utility SaveFig was added in /Utilities to handle figure saving. All
Scripts were modified to use it. All scripts name were containing - were
changed.

Script6_CortexVsGyrus_Fig4cd_Fig5 was edited to report the variance
explained by principal components on the Figure. To this end the utility
PCAAngleTest was also modified.