Adding JunctionPotentialCalculator + minor edits to code
The script JunctionPotentialCaclulator.m was added to the repository.
This script allows to calculate junction potential in our hilar and
cortical interneuron samples based on ion concentration in the intra-
and extra-cellular solutions.

Minor edits were also made to Script6_CortexVsGyrus_Fig4cd_Fig5 so
that it print useful information to a 'Stat.txt' file. This work
hand in hand with minor edits to the function Utilities/PCAAngleTest.m